% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMDadj.R
\name{BMDadjresp}
\alias{BMDadjresp}
\title{Adjusted ED response levels for BMD estimation based on medrc or glsdrc models}
\usage{
BMDadjresp(object, respLev, bmd = c("additional", "extra"),
  background = 0.05)
}
\arguments{
\item{object}{an medrc object}

\item{respLev}{a numeric vector containing the benchmark response levels}

\item{bmd}{benchmark dose estimation (smallest dose resulting in a probability of an abnormal response)}

\item{background}{probability of an abnormal response}
}
\description{
Calculates adjusted response levels for estimation of the BMD
}
\keyword{htest}
