\name{drclogedfunctions}
\alias{LL4.loged}
\alias{LN4.loged}
\alias{W14.loged}
\alias{W24.loged}

\title{ED and BMD estimation on log scale}
\description{Calculate BMD estimates on log transformed scale by the dmlist
  argument in mmaED.}
\usage{
LL4.loged(respLev)
LN4.loged(respLev)
W14.loged(respLev)
W24.loged(respLev)

}
\arguments{
  \item{respLev}{response level}
}

\author{Christian Ritz}

\keyword{models}
