\name{maplot}
\alias{maplot}
\title{medrc average prediction plot}
\description{Plot the average predictions from several medrc objects}
\usage{
maplot(object, ..., ic = "AIC", logx = FALSE)
}
\arguments{
  \item{object}{An object of class medrc}
  \item{\dots}{further objects of class medrc}
  \item{ic}{the information criterion used for model averaging}
  \item{logx}{plot x-axis on logarithmic scale if TRUE}
}

\author{Daniel Gerhard}

\keyword{ hplot }
