% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meAR.R
\name{meAR.2}
\alias{meAR.2}
\alias{meAR.3}
\title{Asymptotic regression model}
\usage{
meAR.2(x, d, e)

meAR.3(x, c, d, e)
}
\arguments{
\item{x}{numeric vector specifying the dose}

\item{d}{upper limit}

\item{e}{\eqn{e>0} is determining the steepness of the increase as \eqn{x}}

\item{c}{lower limit at at \eqn{x=0}}
}
\description{
Providing the mean function and the corresponding self starter function for the asymptotic regression model.
}
\details{
The asymptotic regression model is a three-parameter model with mean function:
\deqn{ f(x) = c + (d-c)(1-\exp(-x/e))}
The parameter \eqn{c} is the lower limit (at \eqn{x=0}), the parameter \eqn{d} is the upper limit and the parameter \eqn{e>0} is determining the steepness of the increase as \eqn{x}.
}
\keyword{models}
