\name{medrc-package}
\alias{medrc-package}
\docType{package}
\title{Mixed effect dose-response curves}
\description{Uses a combination of package drc and nlme to fit nonlinear mixed effects models with ED estimation for the fixed effects.}
\details{
\tabular{ll}{
Package: \tab medrc\cr
Type: \tab Package\cr
Version: \tab 0.0-75\cr
Date: \tab 2016-09-07\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{Daniel Gerhard, Christian Ritz

Maintainer: Daniel Gerhard <gerhard@biostat.uni-hannover.de>
}
\keyword{ package }

