\name{medrcselect}
\alias{medrcselect}
\title{Model selection by information criteria}
\description{Prints several information criteria for several medrc objects}
\usage{
medrcselect(object, ...)
}
\arguments{
  \item{object}{A medrc object}
  \item{\dots}{further medrc objects}
}

\author{Daniel Gerhard}

\keyword{models}
