% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.glsdrc.R
\name{plot.glsdrc}
\alias{plot.glsdrc}
\title{Plot gdrc objects}
\usage{
\method{plot}{glsdrc}(x, ..., ndose = 25, logx = FALSE)
}
\arguments{
\item{x}{An object of class glsdrc}

\item{...}{further arguments}

\item{ndose}{Number of points to interpolate the dose response curve}

\item{logx}{If TRUE, plot x-axis on a logarithmic scale}
}
\description{
Plot predicted dose-response curves for glsdrc objects
}
\keyword{methods}
