\name{plot.glsdrc}
\alias{plot.glsdrc}

\title{Plot gdrc objects}
\description{Plot predicted dose-response curves for glsdrc objects}
\usage{
\method{plot}{glsdrc}(x, ..., ndose=25, logx = FALSE)
}

\arguments{
  \item{x}{An object of class glsdrc}
  \item{ndose}{Number of points to interpolate the dose response curve}
  \item{logx}{If TRUE, plot x-axis on a logarithmic scale}
  \item{...}{further arguments}
}

\author{Daniel Gerhard}



\keyword{ methods }
