% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsdrm.R
\name{glsdrm}
\alias{glsdrm}
\title{Dose-response curve estimation by generalized least squares}
\usage{
glsdrm(form, curveid = NULL, data, fct, correlation = NULL,
  weights = NULL, control = NULL, start = NULL)
}
\arguments{
\item{form}{Formula describing the dose-response relationship}

\item{curveid}{Formula with parameter names on the left hand side (divided by +) and a column name in data, denoting a factor, to estimate separate parameters per factor-level. If NULL only fixed effects for a single curve will be estimated.}

\item{data}{a data.frame object}

\item{fct}{a function of the drc package}

\item{correlation}{additional corClasses object}

\item{weights}{additional varClasses object}

\item{control}{list with nlme control arguments}

\item{start}{optional list with initial values for the fixed components. If NULL the initial values will be found automatically}
}
\value{
An object of class glsdrc
}
\description{
Implements drc nonlinear functions into the nlme framework for nonlinear GLS dose-response modeling.
}
\details{
An application of glsdrm is shown on the help pages of data \code{\link{broccoli}}. EDx and selectivity indices can be calculated with functions \code{\link[drc]{ED}} and \code{\link[drc]{EDcomp}}. Model-averaged ED can be computed by function \code{\link{mmaED}}.
}
\seealso{
\code{\link[drc]{drm}}, \code{\link[nlme]{gls}}
}
\keyword{models}
