% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medrm.R
\name{medrm}
\alias{medrm}
\title{Mixed effects dose-response curves}
\usage{
medrm(form, curveid = NULL, data, fct, random, correlation = NULL,
  weights = NULL, control = NULL, start = NULL, REML = FALSE)
}
\arguments{
\item{form}{Formula describing the dose-response relationship}

\item{curveid}{Formula with parameter names on the left hand side (divided by +) and a column name in data, denoting a factor, to estimate separate parameters per factor-level. If NULL only fixed effects for a single curve will be estimated.}

\item{data}{a data.frame object}

\item{fct}{a function of the drc package}

\item{random}{a list or one-sided formula describing the random effects}

\item{correlation}{additional corClasses object}

\item{weights}{additional varClasses object}

\item{control}{list with nlme control arguments}

\item{start}{optional list with initial values for the fixed components. If NULL the initial values will be found automatically.}

\item{REML}{logical value. If TRUE the model is fit by maximizing the restricted log-likelihood, if FALSE log-likelihood is maximized.}
}
\value{
An object of class medrc
}
\description{
Implements drc nonlinear functions into the nlme framework for mixed effects dose-response modelling.
}
\details{
An application of medrm is shown on the help pages of data \code{\link{broccoli}}. EDx and selectivity indices can be calculated with functions \code{\link[drc]{ED}} and \code{\link[drc]{EDcomp}}. Model-averaged ED can be computed by function \code{\link{mmaED}}.
Marginal EDx and corresponding function are implemented in functions \code{\link{EDmarg}}, using numerical integration conditional on the etimated variance components.
}
\seealso{
\code{\link[drc]{drm}}, \code{\link[nlme]{nlme}}
}
\keyword{models}
