% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meL.R
\name{meL.2}
\alias{meL.2}
\alias{meL.3}
\alias{meL.4}
\alias{meL.5}
\title{The logistic model}
\usage{
meL.2(x, b, e)

meL.3(x, b, d, e)

meL.4(x, b, c, d, e)

meL.5(x, b, c, d, e, f)
}
\arguments{
\item{x}{numeric dose vector.}

\item{b}{steepness}

\item{e}{ED50}

\item{d}{upper limit}

\item{c}{lower limit}

\item{f}{asymmetry}
}
\description{
The general asymmetric five-parameter logistic model for describing dose-response relationships.
}
\details{
The default arguments yields the five-parameter logistic mean function given by the expression \deqn{ f(x) = c + \frac{d-c}{(1+\exp(b(x - e)))^f}} The model is sometimes referred to as the Boltzmann model.
}
\keyword{models}
