% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meLL.R
\name{meLL.2}
\alias{meLL.2}
\alias{meLL.3}
\alias{meLL.4}
\alias{meLL.5}
\title{The log-logistic function}
\usage{
meLL.2(x, b, e)

meLL.3(x, b, d, e)

meLL.4(x, b, c, d, e)

meLL.5(x, b, c, d, e, f)
}
\arguments{
\item{x}{numeric dose vector.}

\item{b}{steepness}

\item{e}{ED50}

\item{d}{upper limit}

\item{c}{lower limit}

\item{f}{asymmetry}
}
\description{
'meLL.X' provides the X-parameter log-logistic function.
}
\details{
The five-parameter logistic function is given by the expression 
\deqn{ f(x) = c + \frac{d-c}{(1+\exp(b(\log(x)-\log(e))))^f}} 
or in another parameterisation 
\deqn{ f(x) = c + \frac{d-c}{(1+\exp(b(\log(x)-e)))^f}} 
The function is asymmetric for \eqn{f} different from 1.
}
\references{
Finney, D. J. (1979) Bioassay and the Practise of Statistical Inference, \emph{Int. Statist. Rev.}, \bold{47}, 1--12.
}
\keyword{models}
