\name{metadrc-methods}
\alias{df.residual.metadrc}
\alias{vcov.metadrc}
\alias{coef.metadrc}
\alias{print.metadrc}
\alias{summary.metadrc}

\title{metadrc methods}
\description{Methods for objects of class metadrc}
\usage{
\method{df.residual}{metadrc}(object, ...)
\method{vcov}{metadrc}(object, ...)
\method{coef}{metadrc}(object, ...)
\method{print}{metadrc}(x, digits = max(3, getOption("digits") - 3), ...)
\method{summary}{metadrc}(object, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{object}{An object of class metadrc}
  \item{x}{An object of class metadrc}
  \item{digits}{An integer giving the number of digits of the parameter coefficients. Default is 3.}
  \item{...}{further arguments}
}

\author{Daniel Gerhard}



\keyword{ methods }
