\name{metadrc-methods}
\alias{df.residual.metadrc}
\alias{vcov.metadrc}

\title{metadrc methods}
\description{Methods for objects of class metadrc}
\usage{
\method{df.residual}{metadrc}(object, ...)
\method{vcov}{metadrc}(object, ...)
}

\arguments{
  \item{object}{An object of class metadrc}
  \item{...}{further arguments}
}

\author{Daniel Gerhard}



\keyword{ methods }
