% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmaBMD.R
\name{mmaBMD}
\alias{mmaBMD}
\title{BMD estimation for averaged medrc or glsdrc models}
\usage{
mmaBMD(object, ..., respLev, ic = c("AIC", "BIC"), interval = c("none",
  "buckland", "kang"), level = 0.95, bmd = c("additional", "extra"),
  background = 0.05, marginal = FALSE, nGQ = 5, rfinterval = c(0, 1000))
}
\arguments{
\item{object}{an medrc object}

\item{...}{further medrc objects}

\item{respLev}{a numeric vector containing the benchmark response levels}

\item{ic}{the information criterion used for model averaging}

\item{interval}{character string specifying the type of confidence intervals to be supplied.}

\item{level}{confidence level, default at 0.95}

\item{bmd}{benchmark dose estimation (smallest dose resulting in a probability of an abnormal response)}

\item{background}{probability of an abnormal response}

\item{marginal}{logical. If TRUE, marginalized BMD estimates are obtained. See BMDmarg for further information.}

\item{nGQ}{integer. Specifies the number nof nodes for Gauss-Hermite quadrature.}

\item{rfinterval}{numeric vector. Interval for root finding (uniroot) to search for ED values.}
}
\description{
Estimates the benchmark dose for averaged fixed effects of several medrc or glsdrc objects
}
\keyword{htest}
